Сделали че-то

+++ PhotosService/Controllers/PhotosApiController.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
@@ -35,11 +36,14 @@ namespace PhotosService.Controllers
         }
 
         [HttpGet("{id}/meta")]
-        public async Task<IActionResult> GetPhotoMeta(Guid id)
+        public async Task<IActionResult> GetPhotoMeta(Guid id, JwtSecurityToken accessToken)
         {
+            
             var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
             if (photoEntity == null)
                 return NotFound();
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
 
             var photo = mapper.Map<PhotoDto>(photoEntity);
             photo.Url = GeneratePhotoUrl(photo);
@@ -47,8 +51,14 @@ namespace PhotosService.Controllers
         }
 
         [HttpGet("{id}/content")]
-        public async Task<IActionResult> GetPhotoContent(Guid id)
+        public async Task<IActionResult> GetPhotoContent(Guid id, JwtSecurityToken accessToken)
         {
+            var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
+            if (photoEntity == null)
+                return NotFound();
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
+            
             var photoContent = await photosRepository.GetPhotoContentAsync(id);
             if (photoContent == null)
                 return NotFound();
@@ -57,8 +67,11 @@ namespace PhotosService.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> AddPhoto(PhotoToAddDto photo)
+        public async Task<IActionResult> AddPhoto(PhotoToAddDto photo, [FromHeader]JwtSecurityToken accessToken)
         {
+            if (accessToken.Subject != photo.OwnerId)
+                return Forbid();
+            
             var content = Convert.FromBase64String(photo.Base64Content);
             var result = await photosRepository.AddPhotoAsync(photo.Title, photo.OwnerId, content);
             if (!result)
@@ -67,12 +80,13 @@ namespace PhotosService.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> UpdatePhoto(Guid id, PhotoToUpdateDto photo)
+        public async Task<IActionResult> UpdatePhoto(Guid id, PhotoToUpdateDto photo, [FromHeader]JwtSecurityToken accessToken)
         {
             var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
             if (photoEntity == null)
                 return NotFound();
-
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
             photoEntity.Title = photo.Title;
             var result = await photosRepository.UpdatePhotoAsync(photoEntity);
             if (!result)
@@ -81,12 +95,14 @@ namespace PhotosService.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeletePhoto(Guid id)
+        public async Task<IActionResult> DeletePhoto(Guid id, JwtSecurityToken accessToken)
         {
             var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
             if (photoEntity == null)
                 return NotFound();
-
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
+            
             var result = await photosRepository.DeletePhotoAsync(photoEntity);
             if (!result)
                 return Conflict();
